Refactor to remove switching logic
