+++ generator/src/test/java/com/scottlogic/deg/generator/walker/reductive/ReductiveRowSpecGeneratorTests.java
@@ -9,6 +9,7 @@ import com.scottlogic.deg.generator.decisiontree.reductive.ReductiveConstraintNo
 import com.scottlogic.deg.generator.fieldspecs.*;
 import com.scottlogic.deg.generator.generation.*;
 import com.scottlogic.deg.generator.generation.databags.RowSpecDataBagSourceFactory;
+import com.scottlogic.deg.generator.generation.databags.StandardRowSpecDataBagSourceFactory;
 import com.scottlogic.deg.generator.reducer.ConstraintReducer;
 import com.scottlogic.deg.generator.restrictions.NullRestrictions;
 import com.scottlogic.deg.generator.restrictions.Nullness;
@@ -60,7 +61,7 @@ class ReductiveRowSpecGeneratorTests {
             GenerationConfig.TreeWalkerType.REDUCTIVE,
             GenerationConfig.CombinationStrategyType.EXHAUSTIVE
         ));
-        RowSpecDataBagSourceFactory dataBagSourceFactory = new RowSpecDataBagSourceFactory(
+        RowSpecDataBagSourceFactory dataBagSourceFactory = new StandardRowSpecDataBagSourceFactory(
             new FieldSpecValueGenerator(
                 config,
                 new StandardFieldValueSourceEvaluator(),
